+++ CHANGES.md
@@ -28,7 +28,9 @@ New Features:
   get_shape and get_connected_shape
 - new function misc.round, implementing the old Python 2 rounding behaviour
 - new class misc.Colour: create colours by names, hex strings or RGB values
-  
+  The class can also handle HSV values. The misc.HSVColour is thus obsolete and 
+  has been removed.
+     
 Changed:
 - major rewrite of stimuli.Shape, bug fixes, see improved documentation 
 - extras
@@ -55,6 +57,7 @@ Changed:
   not a list of tuples. (use misc.geometry.tuples2points to adapt old code)
 - stimuli.Video: attempting to downloaded a missing ffmpeg binary only when
   initializing a Video stimulus with the "mediadecoder" backend
+- misc.HSVColour has been removed and merged with the colour class (misc.colour)
 
 Fixed:
 - Adding field bug in TouchscreenButtonBox